requirements.txt.txt
discord.py
pandas
requests
asyncio
